requirements.txt
aiofiles==0.7.0  # aiofiles
asyncpg==0.23.0  # asyncpg
asynctest==0.13.0  # asynctest
deptree==0.0.10  # deptree
  importlib-metadata==4.6.1  # importlib-metadata
    zipp==3.5.0  # zipp>=0.5
  setuptools==44.0.0  # setuptools
fastapi==0.66.0  # fastapi
  starlette==0.14.2  # starlette==0.14.2
  pydantic==1.8.2  # pydantic!=1.7,!=1.7.1,!=1.7.2,!=1.7.3,!=1.8,!=1.8.1,<2.0.0,>=1.6.2
    typing-extensions==3.10.0.0  # typing-extensions>=3.7.4.3
httpx==0.18.2  # httpx
  sniffio==1.2.0  # sniffio
  httpcore==0.13.6  # httpcore<0.14.0,>=0.13.3
    sniffio==1.2.0  # sniffio==1.*
    anyio==3.2.1  # anyio==3.*
      sniffio==1.2.0  # sniffio>=1.1
      idna==2.10  # idna>=2.8
    h11==0.12.0  # h11<0.13,>=0.11
  rfc3986==1.5.0  # rfc3986[idna2008]<2,>=1.3
  certifi==2021.5.30  # certifi
nltk==3.6.2  # nltk
  tqdm==4.61.2  # tqdm
  regex==2021.7.6  # regex
  click==8.0.1  # click
  joblib==1.0.1  # joblib
pip==20.0.2  # pip
pkg-resources==0.0.0  # pkg-resources
pytest-asyncio==0.15.1  # pytest-asyncio
  pytest==6.2.4  # pytest>=5.4.0
    pluggy==0.13.1  # pluggy<1.0.0a1,>=0.12
    iniconfig==1.1.1  # iniconfig
    py==1.10.0  # py>=1.8.2
    toml==0.10.2  # toml
    packaging==21.0  # packaging
      pyparsing==2.4.7  # pyparsing>=2.0.2
    attrs==21.2.0  # attrs>=19.2.0
python-multipart==0.0.5  # python-multipart
  six==1.16.0  # six>=1.4.0
requests-html==0.10.0  # requests-html
  w3lib==1.22.0  # w3lib
    six==1.16.0  # six>=1.4.1
  bs4==0.0.1  # bs4
    beautifulsoup4==4.9.3  # beautifulsoup4
      soupsieve==2.2.1  # soupsieve>1.2; python_version >= "3.0"
  fake-useragent==0.1.11  # fake-useragent
  requests==2.25.1  # requests
    chardet==4.0.0  # chardet<5,>=3.0.2
    certifi==2021.5.30  # certifi>=2017.4.17
    urllib3==1.26.6  # urllib3<1.27,>=1.21.1
    idna==2.10  # idna<3,>=2.5
  pyquery==1.4.3  # pyquery
    cssselect==1.1.0  # cssselect>0.7.9
    lxml==4.6.3  # lxml>=2.1
  parse==1.19.0  # parse
  pyppeteer==0.2.5  # pyppeteer>=0.0.14
    appdirs==1.4.4  # appdirs<2.0.0,>=1.4.3
    tqdm==4.61.2  # tqdm<5.0.0,>=4.42.1
    pyee==8.1.0  # pyee<9.0.0,>=8.1.0
    urllib3==1.26.6  # urllib3<2.0.0,>=1.25.8
    websockets==8.1  # websockets<9.0,>=8.1
tortoise-orm==0.17.5  # tortoise-orm
  iso8601==0.1.14  # iso8601<0.2.0,>=0.1.13
  aiosqlite==0.16.1  # aiosqlite<0.17.0,>=0.16.0
    typing-extensions==3.10.0.0  # typing_extensions>=3.7.2
  pypika-tortoise==0.1.1  # pypika-tortoise<0.2.0,>=0.1.1
  pytz==2021.1  # pytz
uvicorn==0.14.0  # uvicorn
  asgiref==3.4.1  # asgiref>=3.3.4
  click==8.0.1  # click>=7.*
  h11==0.12.0  # h11>=0.8
wheel==0.36.2  # wheel
